rename stream storage to http stream
