+++ README.md
@@ -3,7 +3,7 @@ Pixels
 
 The core of Pixels is a columnar storage engine designed for data lakes and warehouses.
 It is optimized for analytical tables stored in on-premises and cloud-native storage systems,
-including S3, GCS, HDFS, Redis, and local file systems.
+including S3, GCS, HDFS, Redis, HTTP (Netty), SQS, and local file systems.
 Pixels outperforms Parquet, which is the most widely used columnar format in today's lakehouses, by up to two orders of magnitude.
 
 We have integrated Pixels with popular query engines including DuckDB (1.1.0), Trino (405 and 466), StarRocks (3.3.5), PrestoDB (0.279), and Hive (2.3+).
