+++ pixels-core/pom.xml
@@ -98,7 +98,7 @@
 
         <dependency>
             <groupId>io.pixelsdb</groupId>
-            <artifactId>pixels-storage-stream</artifactId>
+            <artifactId>pixels-storage-http-stream</artifactId>
             <scope>test</scope>
             <optional>true</optional>
         </dependency>
