+++ pixels-storage/pixels-storage-http-stream/src/main/java/io/pixelsdb/pixels/storage/http/PhysicalHttpStreamReader.java
@@ -1,0 +1,162 @@
+/*
+ * Copyright 2024 PixelsDB.
+ *
+ * This file is part of Pixels.
+ *
+ * Pixels is free software: you can redistribute it and/or modify
+ * it under the terms of the Affero GNU General Public License as
+ * published by the Free Software Foundation, either version 3 of
+ * the License, or (at your option) any later version.
+ *
+ * Pixels is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * Affero GNU General Public License for more details.
+ *
+ * You should have received a copy of the Affero GNU General Public
+ * License along with Pixels.  If not, see
+ * <https://www.gnu.org/licenses/>.
+ */
+package io.pixelsdb.pixels.storage.http;
+
+import io.pixelsdb.pixels.common.physical.PhysicalReader;
+import io.pixelsdb.pixels.common.physical.Storage;
+
+import java.io.DataInputStream;
+import java.io.IOException;
+import java.nio.ByteBuffer;
+import java.nio.ByteOrder;
+import java.util.concurrent.CompletableFuture;
+
+import static java.util.Objects.requireNonNull;
+
+public class PhysicalHttpStreamReader implements PhysicalReader
+{
+    private final HttpStream httpStream;
+    private final String path;
+    private final DataInputStream dataInputStream;
+
+    public PhysicalHttpStreamReader(Storage storage, String path) throws IOException
+    {
+        if (storage instanceof HttpStream)
+        {
+            this.httpStream = (HttpStream) storage;
+        }
+        else
+        {
+            throw new IOException("Storage is not LocalFS.");
+        }
+        this.path = path;
+        this.dataInputStream = storage.open(path);
+    }
+
+    @Override
+    public long getFileLength() throws IOException
+    {
+        throw new UnsupportedOperationException("Can't get file length in PhysicalHttpStreamReader");
+    }
+
+    @Override
+    public void seek(long desired) throws IOException
+    {
+        throw new UnsupportedOperationException("Can't seek in PhysicalHttpStreamReader");
+    }
+
+    @Override
+    public ByteBuffer readFully(int length) throws IOException
+    {
+        byte[] buffer = new byte[length];
+        dataInputStream.readFully(buffer);
+        return ByteBuffer.wrap(buffer);
+    }
+
+    @Override
+    public void readFully(byte[] buffer) throws IOException
+    {
+        dataInputStream.readFully(buffer);
+    }
+
+    @Override
+    public void readFully(byte[] buffer, int offset, int length) throws IOException
+    {
+        dataInputStream.readFully(buffer, offset, length);
+    }
+
+    @Override
+    public long readLong(ByteOrder byteOrder) throws IOException
+    {
+        if (requireNonNull(byteOrder).equals(ByteOrder.BIG_ENDIAN))
+        {
+            return dataInputStream.readLong();
+        }
+        else
+        {
+            return Long.reverseBytes(dataInputStream.readLong());
+        }
+    }
+
+    @Override
+    public int readInt(ByteOrder byteOrder) throws IOException
+    {
+        if (requireNonNull(byteOrder).equals(ByteOrder.BIG_ENDIAN))
+        {
+            return dataInputStream.readInt();
+        }
+        else
+        {
+            return Integer.reverseBytes(dataInputStream.readInt());
+        }
+    }
+
+    @Override
+    public boolean supportsAsync() { return false; }
+
+    @Override
+    public CompletableFuture<ByteBuffer> readAsync(long offset, int len) throws IOException
+    {
+        throw new UnsupportedOperationException("Can't read async in PhysicalHttpStreamReader");
+    }
+
+    @Override
+    public void close() throws IOException
+    {
+        this.dataInputStream.close();
+    }
+
+    @Override
+    public String getPath() { return path; }
+
+    @Override
+    public String getPathUri() throws IOException
+    {
+        return httpStream.ensureSchemePrefix(path);
+    }
+
+    /**
+     * Get the port in path.
+     *
+     * @return
+     */
+    @Override
+    public String getName()
+    {
+        if (path == null)
+        {
+            return null;
+        }
+        int slash = path.lastIndexOf(":");
+        return path.substring(slash + 1);
+    }
+
+    @Override
+    public long getBlockId() throws IOException
+    {
+        throw new IOException("Can't get block id in PhysicalHttpStreamReader");
+    }
+
+    @Override
+    public Storage.Scheme getStorageScheme() { return httpStream.getScheme(); }
+
+    @Override
+    public int getNumReadRequests() { return 0; }
+}
