+++ pixels-storage/pixels-storage-http-stream/src/main/resources/META-INF/services/io.pixelsdb.pixels.common.physical.StorageProvider
@@ -1,0 +1,1 @@
+io.pixelsdb.pixels.storage.http.HttpStreamProvider
