+++ pixels-storage/pixels-storage-sqs/src/main/java/io/pixelsdb/pixels/storage/sqs/Main.java
@@ -1,0 +1,13 @@
+package io.pixelsdb.pixels.storage.sqs;
+
+/**
+ * @author hank
+ * @create 2025-09-13
+ */
+public class Main
+{
+    public static void main(String[] args)
+    {
+        System.out.println("Hello, World!");
+    }
+}
