--- pixels-storage/pixels-storage-stream/src/main/resources/META-INF/services/io.pixelsdb.pixels.common.physical.StorageProvider
@@ -1,1 +1,0 @@
-io.pixelsdb.pixels.storage.stream.StreamProvider
