+++ pixels-turbo/pixels-worker-common/src/main/java/io/pixelsdb/pixels/worker/common/StreamWorkerCommon.java
@@ -32,7 +32,6 @@ import org.apache.logging.log4j.Logger;
 import java.io.IOException;
 import java.net.URI;
 import java.util.List;
-import java.util.Objects;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Future;
 import java.util.concurrent.atomic.AtomicReference;
