+++ pixels-turbo/pixels-worker-vhive/pom.xml
@@ -81,7 +81,7 @@
         </dependency>
         <dependency>
             <groupId>io.pixelsdb</groupId>
-            <artifactId>pixels-storage-stream</artifactId>
+            <artifactId>pixels-storage-http-stream</artifactId>
             <optional>true</optional>
         </dependency>
