+++ pixels-turbo/pixels-worker-vhive/src/test/java/TestStreamWorker.java
@@ -49,7 +49,8 @@ import java.util.concurrent.Future;
  * @author jasha64
  * @create 2023-11-03
  */
-public class TestStreamWorker {
+public class TestStreamWorker
+{
     // todo: Has not been updated after integrating worker coordinate service in the workers
 
     static StorageInfo minioEndpoint = new StorageInfo(Storage.Scheme.minio,
