+++ pom.xml
@@ -52,7 +52,8 @@
         <module>pixels-storage/pixels-storage-redis</module>
         <module>pixels-storage/pixels-storage-localfs</module>
         <module>pixels-storage/pixels-storage-mock</module>
-        <module>pixels-storage/pixels-storage-stream</module>
+        <module>pixels-storage/pixels-storage-http-stream</module>
+        <module>pixels-storage/pixels-storage-sqs</module>
         <module>pixels-turbo/pixels-worker-common</module>
         <module>pixels-turbo/pixels-worker-lambda</module>
         <module>pixels-turbo/pixels-invoker-lambda</module>
@@ -201,7 +202,7 @@
             </dependency>
             <dependency>
                 <groupId>io.pixelsdb</groupId>
-                <artifactId>pixels-storage-stream</artifactId>
+                <artifactId>pixels-storage-http-stream</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
